moving JunoMagics import into extensions so that wampify can run without depending on the workings of JunoMagic (eg starting of necessary reactor threads)
